1207
1207

+++ board-management-web/src/main/java/kr/co/ymtech/bm/service/FileExplorerService.java
@@ -19,9 +19,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import kr.co.ymtech.bm.config.PathConfig;
 import kr.co.ymtech.bm.controller.dto.FileDTO;
-import kr.co.ymtech.bm.controller.dto.UploadFileDTO;
+import kr.co.ymtech.bm.controller.dto.FileExplorerDTO;
 import kr.co.ymtech.bm.controller.dto.SaveFolderDTO;
 import kr.co.ymtech.bm.controller.dto.UpdateFileDTO;
+import kr.co.ymtech.bm.controller.dto.UploadFileDTO;
+import kr.co.ymtech.bm.controller.dto.UploadFileResponseDTO;
 
 /**
  * 파일 탐색기 FileExplorerService 클래스
@@ -58,8 +60,9 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 17.
 	 */
 	@Override
-	public List<FileDTO> loadAllFiles(String parentPath, String directoryName) {
-		File file = null;
+	public FileExplorerDTO loadAllFiles(String parentPath, String directoryName) {
+
+		File file;
 
 		if (parentPath == null) {
 			file = new File(PathConfig.getFilePath());
@@ -83,13 +86,16 @@ public class FileExplorerService implements IFileExplorerService {
 					dto.setPath(parentPath);
 
 					if (f.isDirectory()) { // 디렉토리인 경우에만 재귀 호출 수행
-						dto.setChild(loadAllFiles(parentPath, f.getName()));
+						FileExplorerDTO childExplorer = loadAllFiles(parentPath, f.getName());
+						List<FileDTO> childList = childExplorer.getFileList();
+						dto.setChild(childList);
 					}
 
 					list.add(dto);
 				}
 			}
 		}
+
 		list.sort((file1, file2) -> {
 			if (file1.getIsDirectory() && !file2.getIsDirectory()) {
 				return -1; // file1은 디렉토리이고, file2는 파일입니다.
@@ -100,7 +106,11 @@ public class FileExplorerService implements IFileExplorerService {
 			}
 		});
 
-		return list;
+		FileExplorerDTO result = new FileExplorerDTO();
+		result.setFileList(list);
+		result.setFirstPath(PathConfig.getFilePath());
+
+		return result;
 	}
 
 	/**
@@ -113,26 +123,43 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @author 박상현
 	 * @since 2023. 11. 23.
 	 */
-	public void uploadFiles(UploadFileDTO uploadFile) {
+	@Override
+	public UploadFileResponseDTO uploadFiles(UploadFileDTO uploadFile) {
 
-		String directoryPath = null;
-		String filePath = null;
-
-		if ("null".equals(uploadFile.getPath())) {
-			filePath = PathConfig.getFilePath();
+		UploadFileResponseDTO uploadFileResponse = new UploadFileResponseDTO();
+		String directoryPath;
+		String filePath;
+		Integer successCount = 0;
+		Integer failCount = 0;
+		List<String> successFileNames = new ArrayList<String>();
+		
+		if (uploadFile.getPath() == null && uploadFile.getName() == null) {
+			directoryPath = PathConfig.getFilePath();
 		} else {
-			filePath = Paths.get(uploadFile.getPath()).resolve(uploadFile.getName()).normalize().toString();
+			directoryPath = Paths.get(uploadFile.getPath()).resolve(uploadFile.getName()).normalize().toString();
 		}
-
+		
 		for (MultipartFile file : uploadFile.getFiles()) {
-			directoryPath = Paths.get(filePath).resolve(file.getOriginalFilename()).normalize().toString();
+			filePath = Paths.get(directoryPath).resolve(file.getOriginalFilename()).normalize().toString();
 
-			try (InputStream input = file.getInputStream(); OutputStream output = new FileOutputStream(directoryPath)) {
+			File existingFile = new File(filePath);
+			if (existingFile.exists()) {
+				failCount++;
+				continue;
+			}
+			try (InputStream input = file.getInputStream(); OutputStream output = new FileOutputStream(filePath)) {
 				IOUtils.copy(input, output);
+				successCount++;
+				successFileNames.add(file.getOriginalFilename());
 			} catch (IOException e) {
-				System.out.println("파일 업로드 실패");
+				System.out.println("1");
 			}
 		}
+		uploadFileResponse.setFailCount(failCount);
+		uploadFileResponse.setSuccessCount(successCount);
+		uploadFileResponse.setSuccessFileNames(successFileNames);
+		
+		return uploadFileResponse;
 	}
 
 	/**
@@ -149,13 +176,13 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 23.
 	 */
 	@Override
-	public void downloadFile(HttpServletResponse response, String Name, String Path) {
+	public void downloadFile(HttpServletResponse response, String name, String path) {
 
-		String filePath = Paths.get(Path).resolve(Name).normalize().toString();
+		String filePath = Paths.get(path).resolve(name).normalize().toString();
 
 		// 지정된 경로의 폴더에서 파일을 찾아서 다운로드
 		try (FileInputStream input = new FileInputStream(filePath); OutputStream output = response.getOutputStream()) {
-			String fileName = URLEncoder.encode(Name, "UTF-8");
+			String fileName = URLEncoder.encode(name, "UTF-8");
 			fileName = fileName.replaceAll("\\+", "%20");
 			response.setHeader("Content-Disposition", "attachment; filename=\"" + fileName + "\"");
 			IOUtils.copy(input, output);
@@ -177,25 +204,26 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 28.
 	 */
 	@Override
-	public void deleteFile(String Path, String Name) {
+	public String deleteFile(String name, String path) {
 
-		File deletedFile = new File(Paths.get(Path).resolve(Name).normalize().toString());
+		File deletedFile = new File(Paths.get(path).resolve(name).normalize().toString());
 
 		if (deletedFile.isDirectory()) {
 			try {
 				FileUtils.deleteDirectory(deletedFile);
 			} catch (IOException e) {
-				System.out.println("파일 삭제 실패");
+				return "파일 삭제 실패";
 			}
 		} else {
 			deletedFile.delete();
 		}
+		return "파일 삭제 성공!";
 	}
 
 	/**
 	 * @Method saveFolder 서버에 디렉토리를 추가하는 함수
 	 * 
-	 * @param Name          추가할 디렉토리 이름
+	 * @param Name          디렉토리 이름
 	 * @param saveFolderDTO 추가할 디렉토리 정보
 	 * 
 	 * @see kr.co.ymtech.bm.service.IFileExplorerService#saveFolder(java.lang.String,
@@ -205,21 +233,30 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 28.
 	 */
 	@Override
-	public void saveFolder(String Name, SaveFolderDTO saveFolderDTO) {
+	public String saveFolder(String name, SaveFolderDTO saveFolderDTO) {
 
-		String filePath = null;
+		String filePath;
 
-		if ("null".equals(Name)) {
+		if (saveFolderDTO.getPath() == null && name == null) {
 			filePath = Paths.get(PathConfig.getFilePath()).resolve(saveFolderDTO.getNewFolderName()).normalize()
 					.toString();
 		} else {
-			filePath = Paths.get(saveFolderDTO.getPath()).resolve(Name).resolve(saveFolderDTO.getNewFolderName())
+			filePath = Paths.get(saveFolderDTO.getPath()).resolve(name).resolve(saveFolderDTO.getNewFolderName())
 					.normalize().toString();
+		}
+
+		File existingFile = new File(filePath);
+		if (existingFile.exists()) {
+			return "디렉토리 이름중복이므로 생성 실패!";
 		}
 
 		File folder = new File(filePath);
 
-		folder.mkdir();
+		if (folder.mkdir()) {
+			return "디렉토리 생성 성공!";
+		} else {
+			return "디렉토리 생성 실패!";
+		}
 	}
 
 	/**
@@ -233,21 +270,18 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 28.
 	 */
 	@Override
-	public void updateFile(UpdateFileDTO updateFileDTO) {
-		Path oldFileName = null;
-		Path newFileName = null;
+	public String updateFile(UpdateFileDTO updateFileDTO) {
+		Path oldFileName = Paths.get(updateFileDTO.getPath()).resolve(updateFileDTO.getName()).normalize();
+		Path newFileName = Paths.get(updateFileDTO.getPath()).resolve(updateFileDTO.getNewFileName()).normalize();
+		if (Files.exists(newFileName)) {
+			return "파일 이름 변경 실패!";
+		}
 		try {
-			if (updateFileDTO.getPath() == PathConfig.getFilePath()) {
-				oldFileName = Paths.get(PathConfig.getFilePath()).resolve(updateFileDTO.getName()).normalize();
-				newFileName = Paths.get(PathConfig.getFilePath()).resolve(updateFileDTO.getNewFileName()).normalize();
-			} else {
-				oldFileName = Paths.get(updateFileDTO.getPath()).resolve(updateFileDTO.getName()).normalize();
-				newFileName = Paths.get(updateFileDTO.getPath()).resolve(updateFileDTO.getNewFileName()).normalize();
-			}
 			Files.move(oldFileName, newFileName);
 		} catch (IOException e) {
-			System.out.println("파일 이름 변경 실패");
+			return "파일 이름 변경 실패!";
 		}
+		return "파일 이름 변경 성공!";
 	}
 
 	/**
@@ -265,22 +299,26 @@ public class FileExplorerService implements IFileExplorerService {
 	 * @since 2023. 11. 30.
 	 */
 	@Override
-	public void moveFile(String fileName, String folderName, String oldPath, String newPath) {
-		Path oldFilePath;
+	public String moveFile(String fileName, String folderName, String oldPath, String newPath) {
+		Path oldFilePath = Paths.get(oldPath).resolve(fileName).normalize();
 		Path newFilePath;
 		try {
 			if (folderName == null && newPath == null) {
-				oldFilePath = Paths.get(oldPath).resolve(fileName).normalize();
 				newFilePath = Paths.get(PathConfig.getFilePath()).resolve(fileName).normalize();
-
+				if (Files.exists(newFilePath)) {
+					return "파일 이름 중복이므로 이동 실패!";
+				}
 			} else {
-				oldFilePath = Paths.get(oldPath).resolve(fileName).normalize();
 				newFilePath = Paths.get(newPath).resolve(folderName).resolve(fileName).normalize();
+				if (Files.exists(newFilePath)) {
+					return "파일 이름 중복이므로 이동 실패!";
+				}
 			}
 			Files.move(oldFilePath, newFilePath);
 		} catch (IOException e) {
-			System.out.println("파일 이동 실패");
+			return "파일 이동 실패!";
 		}
+		return "파일 이동 성공!";
 	}
 
 }
